Add missing instruction if you are adding typescript
When adding typescript to an existing Create React App project you have to have tsconfig.json in the project, I believe this change should clarify it for beginners getting stuck on this step. Not sure if providing an example would be too much so I have added a link to the tsconfig.json instead.